Revert manager behavior for managers without 'objects'
This fixes reverse many-to-one managers regressions in our project that
were introduced in
https://github.com/typeddjango/django-stubs/commit/665f4d8ea1723d0bcae59904a590ffb03a7c0355.
I've tried to create a reproduce in
https://github.com/typeddjango/django-stubs/pull/300 but was unable to
do so. This patch makes django-stubs work for us again, perhaps you are
able to see why this fix works, because i do not know the reason why my
class doesn't have 'objects'.